Split backend parts into modules, progress on buffered backend
 * mmap and buffered backend moved to extra modules, but users can and
   should import backends from libertem.io.dataset.base
 * Special case for sparse reading is disabled for now, need to evaulate
   if it is worth the effort
 * Added benchmarks for bufered backend, including a sparse ROI case
