+++ benchmarks/io/test_mib.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from libertem import api
 from libertem.udf.masks import ApplyMasksUDF
-from libertem.io.dataset.base.backend import MMapBackend
+from libertem.io.dataset.base import MMapBackend, BufferedBackend
 
 from utils import drop_cache, warmup_cache, get_testdata_prefixes
 
@@ -60,7 +60,11 @@ class TestUseSharedExecutor:
         "prefix", PREFIXES
     )
     @pytest.mark.parametrize(
-        "io_backend", (MMapBackend(enable_readahead_hints=True), None),
+        "io_backend", (
+            MMapBackend(enable_readahead_hints=True),
+            BufferedBackend(),
+            None
+        ),
     )
     def test_mask(self, benchmark, prefix, drop, shared_dist_ctx, io_backend):
         mib_hdr = os.path.join(prefix, MIB_FILE)
@@ -91,6 +95,54 @@ class TestUseSharedExecutor:
             iterations=1
         )
 
+    @pytest.mark.benchmark(
+        group="io"
+    )
+    @pytest.mark.parametrize(
+        "drop", ("cold_cache", "warm_cache")
+    )
+    @pytest.mark.parametrize(
+        "prefix", PREFIXES
+    )
+    @pytest.mark.parametrize(
+        "io_backend", (
+            MMapBackend(enable_readahead_hints=True),
+            BufferedBackend(),
+            None
+        ),
+    )
+    def test_sparse_roi(self, benchmark, prefix, drop, io_backend, shared_dist_ctx):
+        mib_hdr = os.path.join(prefix, MIB_FILE)
+        flist = filelist(mib_hdr)
+
+        ctx = shared_dist_ctx
+        ds = ctx.load(filetype="auto", path=mib_hdr, io_backend=io_backend)
+
+        sparse_roi = np.zeros(ds.shape.nav.size, dtype=np.bool)
+        sparse_roi[::10] = True
+
+        def mask():
+            return np.ones(ds.shape.sig, dtype=bool)
+
+        udf = ApplyMasksUDF(mask_factories=[mask], backends=('numpy', ))
+
+        # warmup executor
+        ctx.run_udf(udf=udf, dataset=ds)
+
+        if drop == "cold_cache":
+            drop_cache(flist)
+        elif drop == "warm_cache":
+            warmup_cache(flist)
+        else:
+            raise ValueError("bad param")
+
+        benchmark.pedantic(
+            ctx.run_udf, kwargs=dict(udf=udf, dataset=ds, roi=sparse_roi),
+            warmup_rounds=0,
+            rounds=1,
+            iterations=1,
+        )
+
 
 @pytest.mark.benchmark(
     group="io"
@@ -102,14 +154,18 @@ class TestUseSharedExecutor:
     "prefix", PREFIXES[:1]
 )
 @pytest.mark.parametrize(
-    "io_backend", (MMapBackend(enable_readahead_hints=True), None),
+    "io_backend", (
+        MMapBackend(enable_readahead_hints=True),
+        BufferedBackend(),
+        None
+    ),
 )
 def test_mask_firstrun(benchmark, prefix, first, io_backend):
     mib_hdr = os.path.join(prefix, MIB_FILE)
     flist = filelist(mib_hdr)
 
     with api.Context() as ctx:
-        ds = ctx.load(filetype="auto", path=mib_hdr)
+        ds = ctx.load(filetype="auto", path=mib_hdr, io_backend=io_backend)
 
         def mask():
             return np.ones(ds.shape.sig, dtype=bool)
@@ -126,7 +182,7 @@ def test_mask_firstrun(benchmark, prefix, first, io_backend):
         warmup_cache(flist)
 
         benchmark.pedantic(
-            ctx.run_udf, kwargs=dict(udf=udf, dataset=ds, io_backend=io_backend),
+            ctx.run_udf, kwargs=dict(udf=udf, dataset=ds),
             warmup_rounds=0,
             rounds=1,
             iterations=1
