+++ src/libertem/io/dataset/base/__init__.py
@@ -1,6 +1,9 @@
 from .exceptions import DataSetException
 from .meta import DataSetMeta, PartitionStructure
 from .roi import _roi_to_indices, _roi_to_nd_indices
+from .backend import LocalFile, IOBackend
+from .backend_mmap import MMapBackend
+from .backend_buffered import BufferedBackend
 from .dataset import DataSet, WritableDataSet
 from .partition import Partition, BasePartition, WritablePartition
 from .utils import FileTree
@@ -9,9 +12,6 @@ from .file import File
 from .tiling import (
     TilingScheme, DataTile, default_get_read_ranges, make_get_read_ranges,
     Negotiator,
-)
-from .backend import (
-    MMapBackend, LocalFile,
 )
 from .decode import (
     Decoder, DtypeConversionDecoder, decode_swap_2, decode_swap_4,
@@ -23,8 +23,9 @@ __all__ = [
     '_roi_to_nd_indices', '_roi_to_indices',
     'DataSet', 'WritableDataSet', 'Partition', 'WritablePartition', 'BasePartition',
     'DataTile', 'FileSet', 'LocalFile', 'File',
-    'FileTree', 'TilingScheme', 'MMapBackend',
+    'FileTree', 'TilingScheme',
     'default_get_read_ranges', 'make_get_read_ranges',
     'Decoder', 'DtypeConversionDecoder', 'Negotiator',
     'decode_swap_2', 'decode_swap_4', 'get_coordinates',
+    'IOBackend', 'BufferedBackend', 'MMapBackend',
 ]
