+++ tests/io/datasets/test_frms6.py
@@ -12,7 +12,7 @@ from libertem.io.dataset.frms6 import (
 from libertem.analysis.raw import PickFrameAnalysis
 from libertem.analysis.sum import SumAnalysis
 from libertem.udf.sumsigudf import SumSigUDF
-from libertem.io.dataset.base import TilingScheme
+from libertem.io.dataset.base import TilingScheme, BufferedBackend
 from libertem.common import Shape
 from libertem.udf.raw import PickUDF
 
@@ -29,6 +29,16 @@ def default_frms6(lt_ctx):
     ds = FRMS6DataSet(path=FRMS6_TESTDATA_PATH)
     ds = ds.initialize(lt_ctx.executor)
     return ds
+
+
+@pytest.fixture
+def buffered_frms6(lt_ctx):
+    buffered = BufferedBackend()
+    return lt_ctx.load(
+        "frms6",
+        path=str(FRMS6_TESTDATA_PATH),
+        io_backend=buffered,
+    )
 
 
 def test_simple_open(default_frms6):
@@ -309,3 +319,29 @@ def test_incorrect_sig_shape(lt_ctx):
     assert e.match(
         r"sig_shape must be of size: 69696"
     )
+
+
+def test_compare_backends(lt_ctx, default_frms6, buffered_frms6):
+    mm_f0 = lt_ctx.run(lt_ctx.create_pick_analysis(
+        dataset=default_frms6,
+        x=0, y=0,
+    )).intensity
+    buffered_f0 = lt_ctx.run(lt_ctx.create_pick_analysis(
+        dataset=buffered_frms6,
+        x=0, y=0,
+    )).intensity
+
+    assert np.allclose(mm_f0, buffered_f0)
+
+
+def test_compare_backends_sparse(lt_ctx, default_frms6, buffered_frms6):
+    roi = np.zeros(default_frms6.shape.nav, dtype=np.bool).reshape((-1,))
+    roi[0] = True
+    roi[1] = True
+    roi[16] = True
+    roi[32] = True
+    roi[-1] = True
+    mm_f0 = lt_ctx.run_udf(dataset=default_frms6, udf=PickUDF(), roi=roi)['intensity']
+    buffered_f0 = lt_ctx.run_udf(dataset=buffered_frms6, udf=PickUDF(), roi=roi)['intensity']
+
+    assert np.allclose(mm_f0, buffered_f0)
